Add checkJs type checking for test helpers

+++ spec/utils/Tester.js
@@ -1,3 +1,4 @@
+// @ts-check
 'use strict';
 
 import { resolve, join } from 'path';
@@ -5,24 +6,52 @@ import { MongoClient } from 'mongodb';
 import execute from './MongoShell.js';
 import JsonValidator from './JsonValidator.js';
 
-const mongodb_port = process.env.MONGODB_PORT || 27017;
-const default_url = `mongodb://localhost:${mongodb_port}/test`;
+/**
+ * @typedef {import('mongodb').Document} MongoDocument
+ * @typedef {import('mongodb').MongoClient} MongoClientType
+ * @typedef {import('mongodb').Collection<MongoDocument>} MongoCollection
+ * @typedef {Record<string, unknown> & { outputFormat?: string }} AnalysisOptions
+ */
+
+const mongodbPort = Number(process.env.MONGODB_PORT || 27017);
+const defaultUrl = `mongodb://localhost:${mongodbPort}/test`;
 
 export default class Tester {
+  /** @type {MongoClientType | undefined} */
+  connection;
+
+  /** @type {MongoCollection | undefined} */
+  coll;
+
+  /**
+   * @param {string} databaseName
+   * @param {string} collectionName
+   */
   constructor(databaseName, collectionName) {
     this.databaseName = databaseName;
     this.collectionName = collectionName;
   }
 
+  /**
+   * @returns {Promise<MongoClientType>}
+   */
   async connect() {
-    const connection = await MongoClient.connect(default_url);
+    const connection = await MongoClient.connect(defaultUrl);
     this.connection = connection;
     this.coll = connection.db(this.databaseName).collection(this.collectionName);
     return connection;
   }
 
+  /**
+   * @param {MongoDocument[]} initialData
+   * @returns {Promise<MongoClientType>}
+   */
   async init(initialData) {
     const connection = await this.connect();
+    if (!this.coll) {
+      throw new Error('Collection connection is not available.');
+    }
+
     await this.coll.deleteMany();
     await this.coll.insertMany(initialData);
     return connection;
@@ -37,7 +66,14 @@ export default class Tester {
     }
   }
 
+  /**
+   * @param {string} dbName
+   */
   getDb(dbName) {
+    if (!this.connection) {
+      throw new Error('MongoDB connection is not available.');
+    }
+
     return this.connection.db(dbName);
   }
 
@@ -45,21 +81,29 @@ export default class Tester {
     return resolve(join(__dirname , '..', '..', 'variety.js'));
   }
 
+  /**
+   * @param {AnalysisOptions} options
+   */
   async runJsonAnalysis(options) {
-    options.outputFormat = 'json';
-    const result = await this.runAnalysis(options, true);
-    return new JsonValidator(JSON.parse(result));
+    const analysisOptions = { ...options, outputFormat: 'json' };
+    const result = await this.runAnalysis(analysisOptions, true);
+    return new JsonValidator(/** @type {JsonValidator['results']} */ (JSON.parse(result)));
   }
 
 
+  /**
+   * @param {AnalysisOptions} [options]
+   * @param {boolean} [quiet=false]
+   */
   runAnalysis(options, quiet) {
-    let str = [];
+    /** @type {string[]} */
+    const str = [];
     if(options) {
-      for(let key in options) {
-        let value = JSON.stringify(options[key]).replace(/"/g, '\'').replace(/\$/g, '\\$');
+      for (const key of Object.keys(options)) {
+        const value = JSON.stringify(options[key]).replace(/"/g, '\'').replace(/\$/g, '\\$');
         str.push(`var ${key}=${value}`);
       }
     }
-    return execute(this.databaseName, null, '"' + str.join(';') + '"', this.getVarietyPath(), quiet, mongodb_port);
+    return execute(this.databaseName, null, str.length > 0 ? '"' + str.join(';') + '"' : undefined, this.getVarietyPath(), quiet, mongodbPort);
   }
 }
